+++ RELEASENOTES.md
@@ -7,6 +7,8 @@ February 24, 2018
 - Converted all `connection` function names and parameter names
   over to `context`. Aliases were implemented for old `connection`
   function and parameter names to reduce possibility of breakage.
+- Added support for creating custom indexing policies when
+  creating a new collection.
 
 ## What is New in CosmosDB 2.0.1